Fixed firebase bug
